move methods not executed by reqmgr out of unifyfs_request_manager
