+++ server/src/unifyfs_fops_rpc.c
@@ -69,16 +69,133 @@ static int rpc_metaset(unifyfs_fops_ctx_t* ctx,
 
 static int rpc_sync(unifyfs_fops_ctx_t* ctx)
 {
-    return rm_cmd_sync_rpc(ctx->app_id, ctx->client_id);
+    size_t i;
+
+    /* assume we'll succeed */
+    int ret = (int)UNIFYFS_SUCCESS;
+
+    /* get memory page size on this machine */
+    int page_sz = getpagesize();
+
+    /* get application client */
+    app_client* client = get_app_client(ctx->app_id, ctx->client_id);
+    if (NULL == client) {
+        return EINVAL;
+    }
+
+    /* get pointer to superblock for this client and app */
+    shm_context* super_ctx = client->shmem_super;
+    if (NULL == super_ctx) {
+        LOGERR("missing client superblock");
+        return EIO;
+    }
+    char* superblk = (char*)(super_ctx->addr);
+
+    /* get pointer to start of key/value region in superblock */
+    char* meta = superblk + client->super_meta_offset;
+
+    /* get number of file extent index values client has for us,
+     * stored as a size_t value in meta region of shared memory */
+    size_t extent_num_entries = *(size_t*)(meta);
+
+    /* indices are stored in the superblock shared memory
+     * created by the client, these are stored as index_t
+     * structs starting one page size offset into meta region
+     *
+     * Is it safe to assume that the index information in this superblock is
+     * not going to be modified by the client while we perform this operation?
+     */
+    char* ptr_extents = meta + page_sz;
+
+    if (extent_num_entries == 0) {
+        return UNIFYFS_SUCCESS;  /* Nothing to do */
+    }
+
+    unifyfs_index_t* meta_payload = (unifyfs_index_t*)(ptr_extents);
+
+    struct extent_tree_node* local_extents = calloc(extent_num_entries,
+                                                    sizeof(*local_extents));
+    if (!local_extents) {
+        LOGERR("failed to allocate memory for local_extents");
+        return ENOMEM;
+    }
+
+    /*
+     * During sync() in client, all index entries in the superblock are
+     * re-written with the segments (unifyfs_rewrite_index_from_seg_tree).
+     * Therefore, the entries here are also ordered by gfid.
+     */
+    int current_count = 0;
+    int current_gfid = meta_payload[0].gfid;
+    struct extent_tree_node* current_extents = local_extents;
+
+    for (i = 0; i < extent_num_entries; i++) {
+        struct extent_tree_node* tmp = &local_extents[i];
+        unifyfs_index_t* meta = &meta_payload[i];
+        int gfid = meta->gfid;
+        size_t offset = meta->file_pos;
+        size_t length = meta->length;
+        size_t logpos = meta->log_pos;
+        unsigned long end = (unsigned long) (offset + length - 1);
+
+        tmp->start = offset;
+        tmp->end = end;
+        tmp->svr_rank = glb_pmi_rank;
+        tmp->app_id = ctx->app_id;
+        tmp->cli_id = ctx->client_id;
+        tmp->pos = (unsigned long) logpos;
+
+        current_count++;
+
+        if (gfid != current_gfid) {
+            ret = unifyfs_inode_add_extents(current_gfid,
+                                            current_count - 1,
+                                            current_extents);
+            if (ret) {
+                LOGERR("failed to add extents (gfid=%d, ret=%d)",
+                        current_gfid, ret);
+                return ret;
+            }
+
+            ret = unifyfs_invoke_broadcast_extents_rpc(current_gfid,
+                       current_count - 1, current_extents);
+            if (ret) {
+                LOGERR("failed to broadcast extents (gfid=%d, ret=%d)",
+                        current_gfid, ret);
+                return ret;
+            }
+
+            current_gfid = gfid;
+            current_extents = tmp;
+            current_count = 1;
+        }
+
+        if (current_count == extent_num_entries) {
+            ret = unifyfs_inode_add_extents(current_gfid,
+                                            current_count,
+                                            current_extents);
+            if (ret) {
+                LOGERR("failed to add extents (gfid=%d, ret=%d)",
+                        current_gfid, ret);
+                return ret;
+            }
+
+            ret = unifyfs_invoke_broadcast_extents_rpc(current_gfid,
+                    current_count, current_extents);
+            if (ret) {
+                LOGERR("failed to broadcast extents (gfid=%d, ret=%d)",
+                        current_gfid, ret);
+                return ret;
+            }
+        }
+    }
+
+    return ret;
 }
 
 static int rpc_fsync(unifyfs_fops_ctx_t* ctx, int gfid)
 {
-    int ret = 0;
-
-    LOGDBG("%s is called but not implemented yet", __func__);
-
-    return ret;
+    return rpc_sync(ctx);
 }
 
 static int rpc_filesize(unifyfs_fops_ctx_t* ctx,
